==> scripts/01-data-prep/04-QSI-Data.R <==
#--------------------------QSI Data Pre-Processing------------------------------
#-Author: A. Jordan Nafa----------------------------Created: February 1, 2022-#
#-R Version: 4.1.2--------------------------------Last Modified: July 8, 2022-#

# Load the necessary libraries for the script----
pacman::p_load(
  "tidyverse",
  "data.table",
  "dtplyr",
  install = FALSE # Set this to true to install packages automatically
)

#-----------------------------------------------------------------------------#
#----------------UCDP Translation Tables and Actor Information-----------------
#-----------------------------------------------------------------------------#

# Load the Dyad ID translation tables
dyad_trans <- read_csv("data/ucdp/translate_dyad.csv", col_types = "cccd") %>% 
  # Exclude one-sided violence dyads
  filter(type_of_violence == 1)  %>% 
  # Split name into side a and side b
  separate(name, into = c("side_a", "side_b"), sep = " - ")

#-----------------------------------------------------------------------------#
#--------------------Importing and Cleaning the QSI Data-----------------------
#-----------------------------------------------------------------------------#

# Read in the QSI data
qsi <- read_csv("data/qsi/RQSI_Replication_data_2021.csv") %>% 
  # Select a subset of the data
  select(conflict, old_id:Gov_Other) %>% 
  # Group the data by group, conflict, and dyad
  group_by(across(c(actorid, DyadID))) %>% 
  # Generate a run-id for conflict
  mutate(conflict_runid = rleid(ArmedConflict), 
         .before = ArmedConflict) %>% 
  # Ungroup the data
  ungroup() %>% 
  # Coerce IDs to character
  mutate(across(actorid:SideB.x, ~ as.character(.x))) %>% 
  # Merge in the UCDP dyad translation table
  left_join(dyad_trans[,1:2], by = c("DyadID" = "old_id")) %>% 
  ## ID for merging with cwsp data
  mutate(
    join_id = case_when(
      new_id == "700" ~ "701",
      new_id == "681" ~ "683",
      new_id == "555" ~ "558",
      TRUE ~ new_id
    ))
  
# Prepare the data for the measurement model, brms requires long form for IRT
qsi_long <- qsi %>% 
  group_by(across(c(join_id, actorid:SideB.x))) %>% 
  # Collapse the data by group, conflict, and dyad taking the maximum 
  summarise(
    ## Start Year for the Conflict Window
    start_year = min(Year),
    ## Stop Year for the Conflict Window 
    stop_year = max(Year),
    ## Institutional Functions
    across(
      c(Pol_Flag:Mil_Other), 
      ~ max(.x, na.rm = TRUE)
    )) %>% 
  # Ungroup the data
  ungroup() %>% 
  # Fix any non-finite values generated by aggregation
  mutate(across(
    Pol_Flag:Mil_Other,
    ~ case_when(
      is.infinite(.x) | is.nan(.x) ~ NA_real_,
      TRUE ~ .x
    )
  )) %>% 
  # Set all of the names to lowercase
  rename_with(
    .fn = ~ str_to_lower(.x), 
    .cols = everything()
    ) %>% 
  # Pivot the data to long form
  pivot_longer(
    cols = pol_flag:mil_other,
    names_to = "item",
    values_to = "qsi"
    ) %>% 
  # Create an identifier for the dimensions
  mutate(dimension = case_when(
    str_detect(item, "pol_") ~ "Political",
    str_detect(item, "econ_") ~ "Economic",
    str_detect(item, "soc_") ~ "Social",
    str_detect(item, "mil_") ~ "Military"
  ))

# Write the Pre-Processed Data to a File
write_rds(
  x = qsi_long, 
  file = "output/project-data/qsi_long_data.rds",
  compress = "gz",
  compression = 9L
)
